requirements.txt
tensorflow
streamlit
numpy
Pillow
scikit-learn
matplotlib
seaborn
pyngrok
opencv-python
